chore(hygiene): remove dead code, dedup crypto-symbol sets (FINLYNQ-113)
CODE_REVIEW.md A6/5 quick wins, all behavior-preserving:

(a) Dead code: delete src/components/google-analytics.tsx (zero importers
    -- the live GA loader is the inline GoogleAnalyticsScripts in
    analytics-consent.tsx) and the dev-gated <NetWorthChart> from the
    dashboard (import + sole render in the devMode block), then remove the
    now-orphaned _components/net-worth-chart.tsx. The correct unconditional
    <NetWorthHistoryChart> stays; netWorthData + useDevMode preserved.

(b) Crypto-symbol dedup: three identical 38-coin CRYPTO_SYMBOLS sets +
    their isCrypto/isCryptoSymbol helpers (holdings-value.ts,
    /api/portfolio/overview, /api/portfolio/crypto) collapse to one canonical
    CRYPTO_HOLDING_SYMBOLS + isCryptoSymbol() in fx/supported-currencies.ts.
    Union check: all three were byte-identical (same 38 coins) so no coin
    needed adding. Also hoisted the two divergent isCurrencyCodeSymbol copies
    to a canonical isCurrencyCodeSymbol(sym, extraCodes?) -- overview passes
    its active_currencies, holdings-value passes nothing, behavior identical.

Distinct from the FX-rate isCryptoCurrency() (BTC/ETH/USDC/USDT), unchanged.

tsc clean; audit:invariants 11/11; lint no new errors; npm test 77 failed /
982 passed / 13 skipped (unchanged baseline).

(Workspace CLAUDE.md stale ignoreBuildErrors:true -> false fixed in 3 places;
that file is untracked/local-only so not in this commit.)
