+++ src/app/(app)/dashboard/page.tsx
@@ -18,7 +18,6 @@ import { OnboardingWizard } from "@/components/onboarding-wizard";
 import { QuickImport } from "./_components/quick-import";
 import { IncomeExpenseChart } from "./_components/income-expense-chart";
 import { SpendingCategoryChart } from "./_components/spending-category-chart";
-import { NetWorthChart } from "./_components/net-worth-chart";
 import { NetWorthHistoryChart } from "@/components/net-worth-history-chart";
 import { AvailableToSpend } from "./_components/available-to-spend";
 import { InsightsSection } from "./_components/insights-section";
@@ -407,7 +406,6 @@ export default function DashboardPage() {
           </div>
           <div className="grid grid-cols-1 lg:grid-cols-3 gap-5">
             <AvailableToSpend income={lastMonthIncome} expenses={lastMonthExpenses} currency={apiDisplayCurrency} />
-            <NetWorthChart data={netWorthData} />
           </div>
           <InsightsSection currency={apiDisplayCurrency} />
         </>
