+++ src/lib/fx/supported-currencies.ts
@@ -44,6 +44,58 @@ export function isMetalCurrency(code: string): boolean {
 }
 
 /**
+ * Crypto ticker symbols recognized as CoinGecko-priced holdings.
+ *
+ * Distinct from `SUPPORTED_CRYPTO_CURRENCIES` / `isCryptoCurrency()`, which
+ * gate the FX rate model (BTC/ETH/USDC/USDT route through CoinGecko as
+ * *currencies*). This wider set classifies a portfolio HOLDING's symbol
+ * (e.g. "SOL", "AVAX", "BTC-USD") as crypto for market-value pricing. It is
+ * the single source of truth for the three portfolio aggregators
+ * (holdings-value.ts, /api/portfolio/overview, /api/portfolio/crypto) that
+ * previously each carried an identical hardcoded copy.
+ *
+ * Adding a coin here makes every aggregator price it consistently; a
+ * divergence used to silently mis-price coins present in only one copy.
+ */
+export const CRYPTO_HOLDING_SYMBOLS = new Set<string>([
+  "BTC", "ETH", "SOL", "ADA", "XRP", "DOGE", "AAVE", "ATOM", "AVAX",
+  "CRV", "FTM", "HBAR", "LINK", "LTC", "MATIC", "POL", "DOT", "XLM",
+  "UNI", "YFI", "SNX", "BNB", "SHIB", "ARB", "OP", "APT", "SUI",
+  "NEAR", "FIL", "ICP", "ALGO", "XTZ", "EOS", "SAND", "MANA", "AXS", "S",
+]);
+
+/**
+ * Classify a holding's symbol as a CoinGecko-priced crypto. Strips the
+ * optional Yahoo-style quote-currency suffix first (e.g. "BTC-USD" -> "BTC").
+ */
+export function isCryptoSymbol(symbol: string | null | undefined): boolean {
+  if (!symbol) return false;
+  return CRYPTO_HOLDING_SYMBOLS.has(symbol.toUpperCase().split("-")[0]);
+}
+
+/**
+ * A holding whose symbol IS a currency code (CAD, USD, EUR, XAU, …)
+ * represents a foreign-cash position, NOT a stock. Yahoo would otherwise
+ * return data for unrelated tickers ("CAD" -> Cadiz Inc on NASDAQ) that
+ * inflate market value by 100×+.
+ *
+ * `extraCodes` lets callers recognize user-defined active currency codes
+ * (e.g. the portfolio overview's `active_currencies` setting) in addition to
+ * the built-in supported set, preserving each call site's exact behavior.
+ */
+export function isCurrencyCodeSymbol(
+  sym: string | null | undefined,
+  extraCodes?: ReadonlySet<string> | readonly string[],
+): boolean {
+  if (!sym) return false;
+  const s = sym.trim().toUpperCase();
+  if (!/^[A-Z]{3,4}$/.test(s)) return false;
+  if (isSupportedCurrency(s)) return true;
+  if (!extraCodes) return false;
+  return Array.isArray(extraCodes) ? extraCodes.includes(s) : (extraCodes as ReadonlySet<string>).has(s);
+}
+
+/**
  * Display labels for common currencies. Falls back to the bare code for anything
  * not listed — keeps the UI legible for the long tail without bundling a full
  * ISO 4217 metadata table.
